requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
